Bug inside DnaSequence
There was a bug that said there was something "Out of Bounds". When there is an out of bounds error that means there will be a bug regarding traversal through a for loop or while loop. So what I did was subtract the length by 2 in the while loop because the dna.charAt(i+1) + dna.charAt(i+2) will go out of bounds. I also threw in a exception as well.